Refactor code into different files
